[update] 메인페이지 파일분리 및 리팩토링 #72
