+++ src/features/main/css/mainFirstSlider.js
