+++ src/features/main/mainpageexample/SecondSliderList.js
@@ -1,0 +1,18 @@
+export const SecondSliderList = [
+  {img : "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png", title:"제목1", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", title:"제목2", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", title:"제목", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://magazine.brique.co/wp-content/uploads/2022/05/전시포스터.jpg", title:"제목3", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png", title:"제목4", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", title:"제목5", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", title:"제목6", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://magazine.brique.co/wp-content/uploads/2022/05/전시포스터.jpg", title:"제목7", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png", title:"제목8", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", title:"제목9", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", title:"제목10", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://magazine.brique.co/wp-content/uploads/2022/05/전시포스터.jpg", title:"제목11", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png", title:"제목12", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", title:"제목13", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", title:"제목14", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"},
+  {img : "https://magazine.brique.co/wp-content/uploads/2022/05/전시포스터.jpg", title:"제목15", date:"2023.03-14-2023.04.20", location:"서울 : 마이아트뮤지엄"}
+]
