+++ src/hooks/main/useEditTime.js
@@ -1,0 +1,6 @@
+export const useEditTime = () => {
+  const editTimehandle = (times) => {
+    return times.split("T")[0].replace("-", ".");
+  };
+  return {editTimehandle}
+}
